Move some deprecated hooks API
