+++ js/hooks/index.js
@@ -1,4 +1,5 @@
 'use strict';
 import './loader';
 
+export * from './deprecated';
 export * from './handlers';
